% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyhtml.R
\name{tidy_html_file}
\alias{tidy_html_file}
\alias{tidy_html_string}
\alias{tidy_html_rd}
\title{Tidy up HTML}
\usage{
tidy_html_file(file)

tidy_html_string(str, filename = "")

tidy_html_rd(file, ...)
}
\arguments{
\item{file}{path to input file}

\item{filename}{(dummy) name to show in output}

\item{...}{passed to \code{tools::Rd2HTML}}
}
\description{
Cleans up HTML and also returns a list of warnings.
}
